moved Channel component and relevant components into a Channel folder
