+++ client/components/Channel/Voting.js
